third_party/blink/web_tests/external/wpt/web-animations/interfaces/KeyframeEffect/target-expected.txt

This is a testharness.js-based test.
[FAIL] Change target from null to a non-existing pseudoElement setting target first.
  assert_equals: Value at 50% progress after setting new target expected "50px" but got "10px"
[FAIL] Change target from null to a non-existing pseudoElement setting pseudoElement first.
  assert_equals: Value at 50% progress after setting new target expected "50px" but got "10px"
[FAIL] Change target from an existing to a different non-existing pseudo-element by setting target.
  assert_equals: Value of 2nd element (currently targeted) after changing the effect target expected "50px" but got "20px"
[FAIL] Change target from an existing to a different non-existing pseudo-element by setting pseudoElement.
  assert_equals: Value of 2nd element (currently targeted) after changing the effect target expected "50px" but got "20px"
[FAIL] Change target from a non-existing to a different non-existing pseudo-element by setting target.
  assert_equals: Value of 2nd element (currently targeted) after changing the effect target expected "50px" but got "20px"
[FAIL] Change target from a non-existing to a different non-existing pseudo-element by setting pseudoElement.
  assert_equals: Value of 2nd element (currently targeted) after changing the effect target expected "50px" but got "20px"
[FAIL] Changing pseudoElement to a non-null invalid pseudo-selector '::abc' throws a SyntaxError
  assert_throws_dom: function "() => effect.pseudoElement = pseudo" did not throw
[FAIL] Changing pseudoElement to ::placeHOLDER works
  assert_equals: expected "::placeholder" but got "::placeHOLDER"
Harness: the test ran to completion.
